refactor(react-query): add commands file for cli actions
